web_backend/requirements.txt
falcon==1.4.1
gunicorn==19.9.0
passlib==1.7
pymongo==3.7.2
mongoengine==0.17.0
ipython==7.3.0
pyyaml==5.1
docopt==0.6.2
gevent==1.4.0
celery==4.2.0
inflection==0.3.1
flower==0.9.3
requests==2.21.0